# openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "binsight",
    "description": "The HTTP API of a binsight server: health, authentication, live events, the tracked wallets, their synchronization and the settings of the instance. Amounts, prices and percentages are canonical decimal strings; every figure carries its exactness.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "1.9.0"
  },
  "paths": {
...
    "/api/v1/positions/{position_id}/events": {
      "get": {
        "tags": [
          "positions"
        ],
        "summary": "Lists a position's movements, newest first.",
        "operationId": "listPositionEvents",
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "description": "The position's permanent id: `<address>-<opening signature>`, both in base58.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "The `next_cursor` of the previous page; absent for the newest movements.",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "How many movements at most (50 by default, 200 at most).",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "currency",
            "in": "query",
            "description": "The currency of the values (`sol` by default).",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/Currency"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A page of movements.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PositionEvents"
                }
              }
            }
          },
          "400": {
            "description": "The id or a query parameter is invalid (`invalid_request`), or the cursor is not one of this timeline (`invalid_cursor`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in (`unauthenticated`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "404": {
            "description": "No tracked wallet holds or held this position (`position_not_found`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "503": {
            "description": "The engine does not serve figures yet (`data_not_ready`).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        },
        "security": [
          {
            "session_cookie": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Currency": {
        "type": "string",
        "description": "The currency figures are shown in.",
        "enum": [
          "sol",
          "usd"
        ]
      },
...
      "DecimalString": {
        "type": "string",
        "description": "An exact decimal number written as text: no exponent, no `+`, no leading zero, no trailing\nzero after the point, never `-0`. For example `\"61.541203117\"`, `\"-0.949\"`, `\"0\"`.",
        "example": "-1.25",
        "pattern": "^(0|-?([1-9][0-9]*(\\.[0-9]*[1-9])?|0\\.[0-9]*[1-9]))$"
      },
...
      "ErrorBody": {
        "type": "object",
        "description": "The body of every error response.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail",
            "description": "What went wrong."
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "Why a request failed. Stable: clients use it as their translation key.",
        "enum": [
          "invalid_request",
          "unauthenticated",
          "invalid_credentials",
          "forbidden_cross_origin",
          "not_found",
          "method_not_allowed",
          "wallet_not_found",
          "position_not_found",
          "request_timeout",
          "invalid_cursor",
          "payload_too_large",
          "too_many_attempts",
          "data_not_ready",
          "internal"
        ]
      },
      "ErrorDetail": {
        "type": "object",
        "description": "The content of an error body.",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "The stable error code; clients translate it."
          },
          "message": {
            "type": "string",
            "description": "An English explanation for debugging; never shown to users as is."
          },
          "request_id": {
            "type": "string",
            "description": "The id of the request, also in the `x-request-id` response header."
          }
        }
      },
      "EventRange": {
        "type": "object",
        "description": "The range a movement set.",
        "required": [
          "lower_bin_id",
          "upper_bin_id"
        ],
        "properties": {
          "lower": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Price",
                "description": "The lower numeric displayed price; the bin ids are the pool's own. Null without a supported quote."
              },
              {
                "type": "null"
              }
            ]
          },
          "lower_bin_id": {
            "type": "integer",
            "format": "int32",
            "description": "The lowest bin."
          },
          "upper": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Price",
                "description": "The upper numeric displayed price; the bin ids are the pool's own. Null without a supported quote."
              },
              {
                "type": "null"
              }
            ]
          },
          "upper_bin_id": {
            "type": "integer",
            "format": "int32",
            "description": "The highest bin."
          }
        }
      },
      "Figure": {
        "oneOf": [
          {
            "type": "object",
            "description": "Exact to the smallest unit.",
            "required": [
              "value",
              "exactness"
            ],
            "properties": {
              "exactness": {
                "type": "string",
                "enum": [
                  "complete"
                ]
              },
              "value": {
                "$ref": "#/components/schemas/Money",
                "description": "The amount."
              }
            }
          },
          {
            "type": "object",
            "description": "A lower bound: some inputs could not be valued and are left out.",
            "required": [
              "value",
              "reasons",
              "exactness"
            ],
            "properties": {
              "exactness": {
                "type": "string",
                "enum": [
                  "partial"
                ]
              },
              "reasons": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Reason"
                },
                "description": "Why it is partial."
              },
              "value": {
                "$ref": "#/components/schemas/Money",
                "description": "The amount of what could be valued."
              }
            }
          },
          {
            "type": "object",
            "description": "Reconstructed rather than measured.",
            "required": [
              "value",
              "reasons",
              "exactness"
            ],
            "properties": {
              "exactness": {
                "type": "string",
                "enum": [
                  "estimated"
                ]
              },
              "reasons": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Reason"
                },
                "description": "Why it is estimated."
              },
              "value": {
                "$ref": "#/components/schemas/Money",
                "description": "The estimated amount."
              }
            }
          },
          {
            "type": "object",
            "description": "Could not be computed.",
            "required": [
              "reasons",
              "exactness"
            ],
            "properties": {
              "exactness": {
                "type": "string",
                "enum": [
                  "unavailable"
                ]
              },
              "reasons": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Reason"
                },
                "description": "Why."
              }
            }
          }
        ],
        "description": "An amount and how far it can be trusted. `value` is present unless the figure is\n`unavailable`; `reasons` say why a figure is not `complete`."
      },
...
      "Money": {
        "type": "object",
        "description": "An exact amount of money.",
        "required": [
          "amount",
          "unit"
        ],
        "properties": {
          "amount": {
            "$ref": "#/components/schemas/DecimalString",
            "description": "The amount in whole units (`\"1.5\"` SOL), signed for gains and losses."
          },
          "unit": {
            "$ref": "#/components/schemas/MoneyUnit",
            "description": "The unit."
          }
        }
      },
      "MoneyUnit": {
        "type": "string",
        "description": "The unit of an amount.",
        "enum": [
          "sol",
          "usd",
          "usdc",
          "usdt"
        ]
      },
      "MovementKind": {
        "type": "string",
        "description": "What a movement did.",
        "enum": [
          "open",
          "add",
          "rebalance_withdrawal",
          "rebalance_deposit",
          "reward_claim",
          "remove",
          "claim",
          "close"
        ]
      },
...
      "PositionEvent": {
        "type": "object",
        "description": "One movement of a position.",
        "required": [
          "at",
          "kind",
          "signature"
        ],
        "properties": {
          "at": {
            "type": "string",
            "format": "date-time",
            "description": "When its transaction happened."
          },
          "base": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DecimalString",
                "description": "The full base token transfer, in whole tokens; `null` for lifecycle or third-mint rewards."
              },
              {
                "type": "null"
              }
            ]
          },
          "kind": {
            "$ref": "#/components/schemas/MovementKind",
            "description": "What it did."
          },
          "price": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Price",
                "description": "The exact bin price of its transaction; `null` when the transaction does not say it."
              },
              {
                "type": "null"
              }
            ]
          },
          "quote": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/DecimalString",
                "description": "The full quote token transfer, in whole tokens; `null` for lifecycle or third-mint rewards."
              },
              {
                "type": "null"
              }
            ]
          },
          "range": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/EventRange",
                "description": "The range it set, when it opened or rebalanced the position."
              },
              {
                "type": "null"
              }
            ]
          },
          "reward": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/RewardMovement",
                "description": "A farming reward paid in its own mint, when this movement is a reward claim."
              },
              {
                "type": "null"
              }
            ]
          },
          "signature": {
            "type": "string",
            "description": "Its transaction signature (base58); several movements can share one."
          },
          "value": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/Figure",
                "description": "The accounting contribution to the header figure, converted at the position's closing\nor spot rate. Rebalance halves contribute only their net addition or withdrawal;\nbase/quote retain their full transfers. Cumulative rounding by category precedes paging.\n`null` for lifecycle events; unavailable when a reward's own price is unknown."
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "PositionEvents": {
        "type": "object",
        "description": "A page of movements, newest first.",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PositionEvent"
            },
            "description": "The movements."
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "description": "The cursor of the next (older) page; `null` on the last page."
          }
        }
      },
...
      "Price": {
        "type": "object",
        "description": "A unit price: quote tokens per base token, with at most twelve significant digits.",
        "required": [
          "amount",
          "quote"
        ],
        "properties": {
          "amount": {
            "$ref": "#/components/schemas/DecimalString",
            "description": "The price."
          },
          "quote": {
            "$ref": "#/components/schemas/PriceQuote",
            "description": "The token it is expressed in."
          }
        }
      },
      "PriceQuote": {
        "type": "string",
        "description": "The token a price is expressed in.",
        "enum": [
          "sol",
          "usdc",
          "usdt"
        ]
      },
...
      "Reason": {
        "oneOf": [
          {
            "type": "object",
            "description": "A wallet holds a token without a price; the figure leaves it out.",
            "required": [
              "wallet",
              "mint",
              "code"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "unpriced_token"
                ]
              },
              "mint": {
                "type": "string",
                "description": "The token's mint."
              },
              "wallet": {
                "type": "string",
                "description": "The wallet that holds it."
              }
            }
          },
          {
            "type": "object",
            "description": "Part of the figure comes from before the wallet was added and is reconstructed.",
            "required": [
              "wallet",
              "until",
              "code"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "reconstructed_history"
                ]
              },
              "until": {
                "type": "string",
                "format": "date-time",
                "description": "When it was added (RFC 3339, UTC): what comes before is reconstructed."
              },
              "wallet": {
                "type": "string",
                "description": "The wallet."
              }
            }
          },
          {
            "type": "object",
            "description": "A wallet's history is still being imported.",
            "required": [
              "wallet",
              "progress",
              "code"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "history_incomplete"
                ]
              },
              "progress": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/DecimalString",
                    "description": "How far the import is, in percent; null when its total is unknown."
                  },
                  {
                    "type": "null"
                  }
                ]
              },
              "wallet": {
                "type": "string",
                "description": "The wallet."
              }
            }
          },
          {
            "type": "object",
            "description": "A position movement, reward, liquidity or fee observation has an unpriced token quantity.",
            "required": [
              "position",
              "code"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "unpriced_leg"
                ]
              },
              "position": {
                "type": "string",
                "description": "The position id."
              }
            }
          },
          {
            "type": "object",
            "description": "A pool is quoted in a token that is neither SOL nor a dollar stablecoin.",
            "required": [
              "pool",
              "code"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "unsupported_quote"
                ]
              },
              "pool": {
                "type": "string",
                "description": "The pool."
              }
            }
          },
          {
            "type": "object",
            "description": "This currency conversion uses the still open UTC day’s rate.",
            "required": [
              "day",
              "code"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "provisional_rate"
                ]
              },
              "day": {
                "type": "string",
                "format": "date",
                "description": "The UTC day that has not closed yet."
              }
            }
          },
          {
            "type": "object",
            "description": "Open PnL is estimated from a mark before the requested historical instant.",
            "required": [
              "age_seconds",
              "code"
            ],
            "properties": {
              "age_seconds": {
                "type": "integer",
                "format": "int64",
                "description": "Whole seconds between the mark and the requested instant.",
                "minimum": 0
              },
              "code": {
                "type": "string",
                "enum": [
                  "stale_mark"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "Positions were open at the requested instant but no preceding valuation is known.",
            "required": [
              "wallet",
              "code"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "missing_open_pnl_mark"
                ]
              },
              "wallet": {
                "type": "string",
                "description": "The wallet whose historical open PnL is unavailable."
              }
            }
          },
          {
            "type": "object",
            "description": "No SOL/USD rate is known for a day the figure needs.",
            "required": [
              "code"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "no_usd_rate"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "A percentage has a zero denominator.",
            "required": [
              "code"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "zero_denominator"
                ]
              }
            }
          },
          {
            "type": "object",
            "description": "A profit factor has no loss to divide by.",
            "required": [
              "code"
            ],
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "no_losses"
                ]
              }
            }
          }
        ],
        "description": "One reason a figure is partial, estimated or unavailable, tagged by `code`."
      },
...
      "RewardMovement": {
        "type": "object",
        "description": "A reward claim in raw units, without assuming the token's decimals or price.",
        "required": [
          "mint",
          "raw_amount",
          "reward_index"
        ],
        "properties": {
          "mint": {
            "type": "string",
            "description": "The mint paid by the pool's reward program, in base58."
          },
          "raw_amount": {
            "type": "string",
            "description": "A canonical integer string of raw token units, not a human token amount.",
            "pattern": "^(0|[1-9][0-9]*)$"
          },
          "reward_index": {
            "type": "integer",
            "format": "int64",
            "description": "Which reward program of the pool paid it.",
            "minimum": 0
          }
        }
      },
...
    },
    "securitySchemes": {
      "session_cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "binsight_session"
      }
    }
  },
...
}